refactor: replace cache with opinionated tool cacher
